94_выносит конвертирование timestamp в строку с датой в отдельные функции расширения корректирует верстку диалога добавляет слушатель нажатий на кнопку confirm в диалоге удаляет лишний переход на выбор фильмов из waitSession
